fastapi-backend/requirements.txt
fastapi
uvicorn
python-dotenv
motor[srv]
pydantic
requests